03-Defensive-Security-and-Blue-Teaming/SigmaToSuricata/requirements.txt
pyyaml
ruamel.yaml
rich
